+++ src/cuemsutils/__init__.py
@@ -1,4 +1,4 @@
 # SPDX-FileCopyrightText: 2025-present Adrià (StageLab) <adria@stagelab.coop>
 #
 # SPDX-License-Identifier: GPL-3.0
-__version__ = "0.1.0rc5"
+__version__ = "0.1.0rc6"